FIX: 앱이 종료된 상태에서 딥링크가 동작하지 않는 문제 수정
