+++ KNUTICE/Utils/NotificationName.swift
@@ -9,6 +9,5 @@ import Foundation
 
 extension Notification.Name {
     static let bookmarkListRefresh = Notification.Name("bookmarkListRefresh")
-    static let pushNotice = Notification.Name("PushNotice")
     static let fcmToken = Notification.Name("FCMToken")
 }
